Add files via upload
It is c program of rock paper scissor game. it is built using the very basic structures like if else loop. I asked the user how many times he want to play the game and based on it the program compute the score of user and the computer
we specially used the structure of rand function which takes the character value in random it is taken by the computer and that character is compared with the user given character after comparison the scores are updated accordingly
In some curious to make the program user friendly we also explained the user how the winner is declared to make the user understand about the game
all my explanations about the game is of my knowledge 
open for any suggestions or any improvements 
 

Signed-off-by: maneesh756 <114976480+maneesh756@users.noreply.github.com>